Move prompt logic to the frontend-core
